+++ changelog.md
@@ -1,6 +1,8 @@
 # Unpublished
 
 + `Breaking` - When `NODE_ENV` is set to `production`, the `faker` and `change` dependencies are excluded. To include them, manually extend jsf via the mocking config object.
++ Added warnings in the place of the stripped `faker` and `chance` modules.
++ Added information about the stripped modules to docs.
 
 # 0.7.5 (June 15th 2016)